Strucutre library and move some code to src
This commit adds a small template to build the library with code
in ./src. The cbps code is reduced to just a small class that can
be refactored to other applications or training mechanisms.

Some things to change are:
 - Figure out regularization. Added a not so fancy L1. Can we
   change the loss?
 - Apply other trainers.
